"fix: add isSuccess checks and make writeTokenDeposit conditional"
Write contract interactions were not waiting for confirmation so would prematurely notify
as successful even when they fail to confirm. Also writeTokenDeposit was being excuted even
when allowance was lower than required amount resulting in a failed transaction which wasn't
supposed to be executed because the allowance was not enough.